main: Use error wrapper to make workflow more concise
channel: Remove xdr.SorobanAuthorizationEntry since the authorization is completed by stellar/go since the recent versions. Also remove makePreImgAuth, which was used to calculate the Signature field in the SorobanAuthorizationEntry. This is now obsolete.

channel/env: Remove the TestInteractContract function, which was used to invoke the testinteract function of the payment channel contract. This was used for testing purposes only and is now obsolete.
Include a global shared mutex, sharedMtx, to prevent both users invoking the same contract function at the same time.

util: Remove CreateFundNewRandomStellarKP, which is now substituted by CreateFundStellarAccounts. Remove SorobanAuthorizationEntry and the i128Param function because it is unused.

testdata: perun_soroban_contract and perun_soroban_token are updated to represent the most recent versions of the contracts.